refactor(utxo-lib): move `getChainAndIndexFromPath` to separate file
Move chain and index parsing function to ScriptId.ts in wallet module for
improved code organization. Add dedicated test file and improved validation.

Issue: BTC-1966
