+++ modules/utxo-lib/src/bitgo/wallet/index.ts
@@ -7,3 +7,4 @@ export * from './WalletScripts';
 export * from './WalletKeys';
 export * from './psbt/PsbtOutputs';
 export * from './psbt/RootNodes';
+export * from './ScriptId';
